chore(weave): Move trace tests into 1 dir (#2523)
