+++ weave/tests/test_patcher_module/example_class.py
