+++ weave/tests/trace/test_patcher.py
@@ -6,11 +6,11 @@ from weave.trace.patcher import SymbolPatcher
 
 
 def test_symbol_patcher():
-    from .test_patcher_module.example_class import ExampleClass
+    from weave.tests.test_patcher_module.example_class import ExampleClass
 
     patcher = SymbolPatcher(
         lambda: importlib.import_module(
-            "weave.trace.tests.test_patcher_module.example_class"
+            "weave.tests.test_patcher_module.example_class"
         ),
         "ExampleClass.example_fn",
         lambda original_fn: lambda self: 43,
@@ -40,7 +40,7 @@ def test_symbol_patcher_invalid_module():
 def test_symbol_patcher_invalid_attr():
     patcher = SymbolPatcher(
         lambda: importlib.import_module(
-            "weave.trace.tests.test_patcher_module.example_class"
+            "weave.tests.test_patcher_module.example_class"
         ),
         "NotARealExampleClass.example_fn",
         lambda original_fn: lambda self: 43,
@@ -52,11 +52,11 @@ def test_symbol_patcher_invalid_attr():
 
 @pytest.mark.disable_logging_error_check
 def test_symbol_patcher_invalid_patching(log_collector):
-    from .test_patcher_module.example_class import ExampleClass
+    from weave.tests.test_patcher_module.example_class import ExampleClass
 
     patcher = SymbolPatcher(
         lambda: importlib.import_module(
-            "weave.trace.tests.test_patcher_module.example_class"
+            "weave.tests.test_patcher_module.example_class"
         ),
         "ExampleClass.example_fn",
         lambda original_fn: [] + 42,
