+++ weave/tests/trace/test_weave_client.py
@@ -11,6 +11,7 @@ import requests
 import weave
 import weave.trace_server.trace_server_interface as tsi
 from weave import Evaluation
+from weave.tests.trace.testutil import ObjectRefStrMatcher
 from weave.tests.trace.util import (
     AnyIntMatcher,
     DatetimeMatcher,
@@ -27,7 +28,6 @@ from weave.trace.refs import (
     TABLE_ROW_ID_EDGE_NAME,
 )
 from weave.trace.serializer import get_serializer_for_obj, register_serializer
-from weave.trace.tests.testutil import ObjectRefStrMatcher
 from weave.trace_server.sqlite_trace_server import SqliteTraceServer
 from weave.trace_server.trace_server_interface import (
     FileContentReadReq,
