+++ weave/tests/trace/test_weave_object_vals.py
