+++ weave/tests/trace/testutil.py
